requirements.txt
fastapi
uvicorn
sqlalchemy
stripe
python-dotenv
pydantic